Edited day 0 log
Changed the Day 0 date, and the project name. I filled out the progess part. I changed the thoughts part. I added the Improve part, to add a start stop and repeat section so as i now not to make the same mistakes as yesterday, and to start doing what i missed out on today. To repeat ding what i did good this day.